BiometricViewModel created
